Add new jobTitle function to name object.
I had to make a few concessions when doing this.
Ideally I would have added this function under the 'company' module,
but I chose to add it to the 'name' object instead,
as the data happened to already exist in the locale files under
'name.title'.

However, I felt that the field names in the locale files were confusing,
so their wrappers in the name module are renamed for clarity.
For example, faker.name.jobArea() returns a random element from
locale.name.title.level. In my opinion, 'jobArea' was a much better
description of the data than 'level'.

Finally, fixed a bug in one unit test which didn't restore a function
to its original state after watching it.