Bugfix random.shuffle last element.
There was a bug in the implementation of the modern Fisher–Yates
algorithm in that `i` was being decremented prematurely in each loop.
This caused the last element of the array to never be swapped
and therefore remain the last element.

Updated doc to denote the provided array is shuffled in place.