feat(person): add job titles for fr
This uses the new job_title_pattern to add job titles with a logical order for French e.g.

```
Coordinateur en chef de la tactique
Executif central des facteurs
Stagiaire corporate des marchés
```

Due to the need to translate “of” in French to de, de la, des, du, etc depending on the number and gender of the noun, this is incorporated into the level definitions.

The fr_BE and fe_CH definitions are deleted so the standard fr ones can fall through.